requirements.txt
gradio
openai
python-dotenv
requests
beautifulsoup4
